Day 6/rag-authorship-checker/requirements.txt
streamlit
pymupdf
google-generativeai
pandas
requests
